Move loggers to separate implementation file
