+++ src/pch.h
@@ -21,5 +21,6 @@
 
 #include "utils/binaryio.h"
 #include "utils/utils.h"
+#include "utils/logger.h"
 
 #include "application/repak.h"
