+++ src/utils/utils.cpp
@@ -63,51 +63,3 @@ std::string Utils::ChangeExtension(const std::string& in, const std::string& ext
 {
 	return std::filesystem::path(in).replace_extension(ext).u8string();
 }
-
-void Warning(const char* fmt, ...)
-{
-	va_list args;
-	va_start(args, fmt);
-
-	std::string msg = "WARNING: " + std::string(fmt);
-
-	vprintf(msg.c_str(), args);
-
-	va_end(args);
-}
-
-void Error(const char* fmt, ...)
-{
-	va_list args;
-	va_start(args, fmt);
-
-	std::string msg = "ERROR: " + std::string(fmt);
-
-	vprintf(msg.c_str(), args);
-
-	va_end(args);
-
-	exit(EXIT_FAILURE);
-}
-
-void Log(const char* fmt, ...)
-{
-	va_list args;
-	va_start(args, fmt);
-	vprintf(fmt, args);
-	va_end(args);
-}
-
-void Debug(const char* fmt, ...)
-{
-#ifdef _DEBUG
-	va_list args;
-	va_start(args, fmt);
-
-	std::string msg = "[D] " + std::string(fmt);
-
-	vprintf(msg.c_str(), args);
-
-	va_end(args);
-#endif
-}
